requirements.txt
Jinja2==3.1.6
MarkupSafe==3.0.2
mypy==1.17.1
mypy_extensions==1.1.0
pathspec==0.12.1
PyYAML==6.0.2
typing_extensions==4.15.0
